Reworking file organization to better separate into packages.
Adding in __init__.py for WUBRG.
