+++ WUBRG/__init__.py
@@ -1,0 +1,2 @@
+from consts import *
+from funcs import *
