+++ game_metadata/Card.py
@@ -2,7 +2,7 @@ from typing import Union
 from enum import Flag, auto
 
 from data_handling.utils.consts import RARITY_ALIASES
-from WUBRG.WUBRG import get_color_identity
+from WUBRG import get_color_identity
 
 
 # https://scryfall.com/docs/api/layouts
