+++ game_metadata/FormatMetadata.py
@@ -2,7 +2,7 @@ from typing import Union, Optional
 from functools import cmp_to_key
 from datetime import date, timedelta
 
-from WUBRG.WUBRG import COLOR_INDEXES
+from WUBRG import COLOR_INDEXES
 from utils.settings import SETS, FORMATS, SET_CONFIG
 from game_metadata.CallScryfall import CallScryfall
